feature: se cambia la estructura para hacerla mas modular y se implementa el cors.
