+++ .gitignore
@@ -1,0 +1,56 @@
+# Python
+_pycache__/
+*.pyc
+*.pyo
+*.pyd
+crud.cpython-313.pyc
+database.cpython-313.pyc
+main.cpython-313.pyc
+schemas.cpython-313.pyc
+models.cpython-313.pyc
+untils.cpython-313.pyc
+
+# Node.js
+node_modules/
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Environment variables
+.env
+.env.local
+.env.*.local
+venv/
+env/
+.venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Build output
+/dist
+/build
+/.next
+/out
+
+# IDEs and editors
+.vscode/
+.idea/
+*.swp
+*.sublime-workspace
+*.sublime-project
+
+# OS files
+.DS_Store
+Thumbs.db
+
+# Logs
+logs/
+*.log
+
+# Coverage
+coverage/
+
+# Misc
+*.bak
+*.tmp
