+++ database.py
@@ -1,7 +1,16 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 
+# Configuración de la base de datos
 DATABASE_URL = "mssql+pyodbc://@localhost\SQLEXPRESS/CentroSalud?driver=ODBC+Driver+17+for+SQL+Server&trusted_connection=yes"
 engine = create_engine(DATABASE_URL)
 Base = declarative_base()
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+# Dependencia para obtener la sesión de la base de datos
+def get_db():
+    db = SessionLocal() #crea sesion
+    try:  #entrega la sesion
+        yield db
+    finally: #termina la sesion
+        db.close()
